Add ES.ToLength helper.
The ES6 spec added a ToLength helper in the August 23, 2013 draft (rev 17).
This is used to eliminate the 2^32 length limit for array-like
objects.  Audit places in the ES6 spec where ToLength is used and
ensure we do the same in es6-shim.  Some additional minor
spec-compliance fixups were made in the process.