Update chai configuration.
The old `chai.Assertion...` style has been deprecated in chai 1.9.1.
What's more, it triggers warnings when used, and those warnings use
`console.warn`... which isn't supported on IE <= 8.  So this was causing
testling failures on IE 8 which weren't even our fault.

This should return us to all-green on testling.

We also bumped the version of the promises-aplus-tests dependency
while we were at it.  That synchronizes the version of mocha
expected for promises-aplus-tests with the version we expect.