Fix compatibility bugs with Firefox 25, 30, 31.
Shim `Array#find` and `Array#findIndex` if they don't match ES6 spec's
treatment of holes.  (Be sure to do this test after `Number` has been
shimmed, and we've already eliminated a premature use of `Math.sign`.)

Broaden the permissible values of `Math.cosh` (Firefox 31 on Linux/x86
shows a difference here.)

Fix an overly-specific test for `Symbol.iterator` (for compatibility,
we use `'@@iterator'` as the name of our iterator on Firefox, in the
absence of `Symbol.iterator`).

The `defineProperties` method doesn't work if we are using
`Symbol.iterator` as the name for the iterator method.  Only newer
browsers define enough of `Symbol` to encounter this problem.  Be
careful when shimming iterators.