Fix an IE bug where the layout engine internally calls the userland `Object.getOwnPropertyNames`.
This is nasty, but since this only occurs when:
a) an iframe is created, in which case, it will be a fresh `window`, so the `cachedWindowNames` should be the same list
b) a user manually grabs the `window` object from a different realm (ie, cross-frame iframe), and passes it into `Object.getOwnPropertyNames`. I'm not sure if this is even possible, but since trying to *do* anything with the property names on the foreign `window` object will throw a security error, I'm OK with not throwing one here.

Fixes #333