Update Promise implementation to match latest ES6 spec.
Includes the erratum for Promise.resolve discussed in #344.

Also fixes the Map/Set constructors, since the spec uses [[Construct]]
now instead of a hidden create method, and the previous version
misapplied `Symbol.species` in this role.

Also added ES.IteratorClose and updated a number of places where that
is used in the latest spec, including in the implementation of
`Array.from`.

Closes: #239
Closes: #344