Fix for IE11 when startsWith String prototype method doesn't exist
When in IE11 the 'preserve string' method crashes on `startsWith`. `startsWith` doesn't exist in IE11 so the `preserveToString` method doesn't work (there's nothing to preserve).

startsWith compliant check also checks if the method exists, and overrideNative throws if the original property isn't present

Added curly braces to pass linter

Rework around the way 'startsWith' is handled in IE

Fixes for the linter

Made code review edits

README tweaks.

Fix for IE11 check before `preserveToString`

When in IE11 the 'preserve string' method crashes on `startsWith`. `startsWith` doesn't exist in IE11 so the `preserveToString` method doesn't work (there's nothing to preserve).

Added a check to prevent it running if it doesn't need to and it's going to cause an error.

Added curly braces to pass linter

Fixes for the linter

A slight rework around handling  not being present in IE