Ensure that ES.* abstract operations are able to be inlined.
Add a bogus assignment to a function prototype, which appears to be the
magic hint to treat function assignments as constant (until proven
otherwise).  Also get rid of the after-the-fact mutation of the ES
object by making IsPromise an ordinary function.  (We could also
call `Object.freeze(ES)` but it doesn't actually improve performance
further.)